requirements.txt
numpy
flask==3.1.0
pillow==11.1.0
torch==2.5.1
torchvision==0.20.1
gunicorn==23.0.0
spacy==3.8.4
pandas==2.2.3
Werkzeug==3.1.3


